refactor: Migrate Label-related States to label.state
Transition all states pertinent to the label to /label/label.state,
enhancing clarity and modularity of the state management process.
