+++ components/ui/comboBoxes/labelComboBox.tsx
@@ -8,13 +8,13 @@ import { useLabelChangeHandler, useDataButtonComboboxFilterLabel } from '@hooks/
 import { useLabelModalStateOpen } from '@hooks/modals';
 import { Labels, Types } from '@lib/types';
 import { atomComboBoxQuery } from '@states/comboBoxes';
-import { selectorSelectedLabels, selectorComboBoxFilteredLabels } from '@states/labels';
 import { useRecoilValue } from 'recoil';
 import { ComboBox } from './comboBox';
 import { ComboBoxNewItemButton } from './comboBox/comboBoxNewItemButton';
 import { Fragment as LabelComboBoxFragment } from 'react';
 import { classNames } from '@stateLogics/utils';
 import { SvgIcon } from '@icon/svgIcon';
+import { selectorSelectedLabels, selectorComboBoxFilteredLabels } from '@label/label.states';
 
 type Props = Partial<Pick<Types, 'todo' | 'selectedQueryLabels'>>;
