+++ components/ui/dropdowns/v1/labelComboBoxDropdown.tsx
@@ -1,7 +1,6 @@
 import { PrefetchRouterButton } from '@buttons/button/prefetchRouterButton';
 import { IconButton } from '@buttons/iconButton';
 import { Types } from '@lib/types';
-import { selectorSelectedLabels } from '@states/labels';
 import { LabelComboBox } from '@ui/comboBoxes/labelComboBox';
 import { LabelsHorizontalGradients } from '@ui/gradients/labelsHorizontalGradients';
 import { Fragment as LabelComboBoxDropdownFragment, useRef } from 'react';
@@ -17,6 +16,7 @@ import { useTodoModalStateClose } from '@hooks/modals';
 import { selectorSessionTodoItem } from '@states/atomEffects/todos';
 import { atomTodoNew } from '@states/todos';
 import { classNames, paths } from '@stateLogics/utils';
+import { selectorSelectedLabels } from '@label/label.states';
 
 type Props = Partial<Pick<Types, 'selectedQueryLabels' | 'container' | 'todo'>>;
